+++ development/tools/jobs/governance/test-infra_governance_test.go
@@ -1,4 +1,4 @@
-package addons_test
+package governance_test
 
 import (
 	"fmt"
@@ -10,17 +10,16 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestAddonsGovernanceJobPresubmit(t *testing.T) {
+func TestTestInfraGovernanceJobPresubmit(t *testing.T) {
 	// WHEN
-	jobConfig, err := tester.ReadJobConfig("./../../../../prow/jobs/addons/addons-governance.yaml")
+	jobConfig, err := tester.ReadJobConfig("./../../../../prow/jobs/governance.yaml")
 	// THEN
 	require.NoError(t, err)
 
-	assert.Len(t, jobConfig.PresubmitsStatic, 1)
-	presubmits := jobConfig.AllStaticPresubmits([]string{"kyma-project/addons"})
+	presubmits := jobConfig.AllStaticPresubmits([]string{"kyma-project/test-infra"})
 	assert.Len(t, presubmits, 1)
 
-	expName := "pre-master-addons-governance"
+	expName := "pre-master-test-infra-governance"
 	actualPresubmit := tester.FindPresubmitJobByNameAndBranch(presubmits, expName, "master")
 	require.NotNil(t, actualPresubmit)
 	assert.Equal(t, expName, actualPresubmit.Name)
@@ -28,36 +27,33 @@ func TestAddonsGovernanceJobPresubmit(t *testing.T) {
 	assert.Equal(t, 10, actualPresubmit.MaxConcurrency)
 	assert.False(t, actualPresubmit.SkipReport)
 	assert.True(t, actualPresubmit.Decorate)
-	assert.Equal(t, "github.com/kyma-project/addons", actualPresubmit.PathAlias)
-	tester.AssertThatHasExtraRefTestInfra(t, actualPresubmit.JobBase.UtilityConfig, "master")
 	tester.AssertThatHasPresets(t, actualPresubmit.JobBase, preset.BuildPr, preset.DindEnabled)
 	assert.Equal(t, "milv.config.yaml|.md$", actualPresubmit.RunIfChanged)
 	assert.True(t, tester.IfPresubmitShouldRunAgainstChanges(*actualPresubmit, true, "milv.config.yaml"))
 	assert.True(t, tester.IfPresubmitShouldRunAgainstChanges(*actualPresubmit, true, "some_markdown.md"))
-	assert.Equal(t, tester.ImageBootstrapLatest, actualPresubmit.Spec.Containers[0].Image)
+	assert.Equal(t, tester.ImageBootstrapTestInfraLatest, actualPresubmit.Spec.Containers[0].Image)
 	assert.Equal(t, []string{tester.GovernanceScriptDir}, actualPresubmit.Spec.Containers[0].Command)
-	assert.Equal(t, []string{"--repository", "addons"}, actualPresubmit.Spec.Containers[0].Args)
+	assert.Equal(t, []string{"--repository", "test-infra"}, actualPresubmit.Spec.Containers[0].Args)
 }
 
-func TestAddonsGovernanceJobPeriodic(t *testing.T) {
+func TestTestInfraGovernanceJobPeriodic(t *testing.T) {
 	// WHEN
-	jobConfig, err := tester.ReadJobConfig("./../../../../prow/jobs/addons/addons-governance.yaml")
+	jobConfig, err := tester.ReadJobConfig("./../../../../prow/jobs/governance.yaml")
 	// THEN
 	require.NoError(t, err)
 
 	periodics := jobConfig.AllPeriodics()
-	assert.Len(t, periodics, 1)
 
-	expName := "addons-governance-nightly"
+	expName := "test-infra-governance-nightly"
 	actualPeriodic := tester.FindPeriodicJobByName(periodics, expName)
 	require.NotNil(t, actualPeriodic)
 	assert.Equal(t, expName, actualPeriodic.Name)
 	assert.True(t, actualPeriodic.Decorate)
-	assert.Equal(t, "0 22 * * 1-5", actualPeriodic.Cron)
+	assert.Equal(t, "0 5 * * 1-5", actualPeriodic.Cron)
 	tester.AssertThatHasPresets(t, actualPeriodic.JobBase, preset.DindEnabled)
-	tester.AssertThatHasExtraRepoRef(t, actualPeriodic.JobBase.UtilityConfig, []string{"test-infra", "addons"})
-	assert.Equal(t, tester.ImageBootstrapLatest, actualPeriodic.Spec.Containers[0].Image)
+	tester.AssertThatHasExtraRefTestInfra(t, actualPeriodic.JobBase.UtilityConfig, "master")
+	assert.Equal(t, tester.ImageBootstrapTestInfraLatest, actualPeriodic.Spec.Containers[0].Image)
 	assert.Equal(t, []string{tester.GovernanceScriptDir}, actualPeriodic.Spec.Containers[0].Command)
-	repositoryDirArg := fmt.Sprintf("%s/addons", tester.KymaProjectDir)
-	assert.Equal(t, []string{"--repository", "addons", "--repository-dir", repositoryDirArg, "--full-validation", "true"}, actualPeriodic.Spec.Containers[0].Args)
+	repositoryDirArg := fmt.Sprintf("%s/test-infra", tester.KymaProjectDir)
+	assert.Equal(t, []string{"--repository", "test-infra", "--repository-dir", repositoryDirArg, "--full-validation", "true"}, actualPeriodic.Spec.Containers[0].Args)
 }
