--- prow/jobs/addons/addons-governance.yaml
@@ -1,70 +1,0 @@
-presubmits: # runs on PRs
-  kyma-project/addons:
-    - name: pre-master-addons-governance
-      cluster: untrusted-workload
-      decorate: true
-      max_concurrency: 10
-      branches:
-        - ^master$
-        - ^main$
-      labels:
-        preset-dind-enabled: "true"
-        preset-build-pr: "true"
-      run_if_changed: "milv.config.yaml|.md$"
-      path_alias: github.com/kyma-project/addons
-      extra_refs:
-        - org: kyma-project
-          repo: test-infra
-          base_ref: master
-          path_alias: github.com/kyma-project/test-infra
-      spec:
-        containers:
-          - image: eu.gcr.io/kyma-project/prow/test-infra/bootstrap:v20181121-f3ea5ce
-            securityContext:
-              privileged: true
-            command:
-              - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/governance.sh"
-            args:
-              - "--repository"
-              - "addons"
-            resources:
-              requests:
-                memory: 10Mi
-                cpu: 100m
-
-periodics: # runs periodic on master
-- name: addons-governance-nightly
-  cluster: untrusted-workload
-  annotations:
-    testgrid-create-test-group: "false"
-  decorate: true
-  cron: "0 22 * * 1-5" # “At 01:00 on every day-of-week from Monday through Friday.”
-  labels:
-    preset-dind-enabled: "true"
-  extra_refs:
-    - org: kyma-project
-      repo: test-infra
-      base_ref: master
-      path_alias: github.com/kyma-project/test-infra
-    - org: kyma-project
-      repo: addons
-      base_ref: master
-      path_alias: github.com/kyma-project/addons
-  spec:
-    containers:
-      - image: eu.gcr.io/kyma-project/prow/test-infra/bootstrap:v20181121-f3ea5ce
-        securityContext:
-          privileged: true
-        command:
-          - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/governance.sh"
-        args:
-          - "--repository"
-          - "addons"
-          - "--repository-dir"
-          - "/home/prow/go/src/github.com/kyma-project/addons"
-          - "--full-validation"
-          - "true"
-        resources:
-          requests:
-            memory: 10Mi
-            cpu: 100m
