--- prow/jobs/test-infra/test-infra-governance.yaml
@@ -1,66 +1,0 @@
-# Code generated by rendertemplates. DO NOT EDIT.
-
-
-presubmits: # runs on PRs
-  kyma-project/test-infra:
-    - name: pre-master-test-infra-governance
-      labels:
-        preset-build-pr: "true"
-        preset-dind-enabled: "true"
-      run_if_changed: 'milv.config.yaml|.md$'
-      skip_report: false
-      decorate: true
-      cluster: untrusted-workload
-      max_concurrency: 10
-      branches:
-        - ^master$
-        - ^main$
-      spec:
-        containers:
-          - image: "eu.gcr.io/kyma-project/test-infra/bootstrap:PR-2951"
-            securityContext:
-              privileged: true
-            command:
-              - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/governance.sh"
-            args:
-              - "--repository"
-              - "test-infra"
-            resources:
-              requests:
-                memory: 10Mi
-                cpu: 100m
-  
-periodics: # runs on schedule
-    - name: test-infra-governance-nightly
-      annotations:
-        testgrid-create-test-group: "false"
-      labels:
-        preset-dind-enabled: "true"
-      cron: "0 5 * * 1-5"
-      skip_report: false
-      decorate: true
-      cluster: untrusted-workload
-      extra_refs:
-        - org: kyma-project
-          repo: test-infra
-          path_alias: github.com/kyma-project/test-infra
-          base_ref: master
-      spec:
-        containers:
-          - image: "eu.gcr.io/kyma-project/test-infra/bootstrap:PR-2951"
-            securityContext:
-              privileged: true
-            command:
-              - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/governance.sh"
-            args:
-              - "--repository"
-              - "test-infra"
-              - "--repository-dir"
-              - "/home/prow/go/src/github.com/kyma-project/test-infra"
-              - "--full-validation"
-              - "true"
-            resources:
-              requests:
-                memory: 10Mi
-                cpu: 100m
-
