--- templates/data/test-infra-governance-data.yaml
@@ -1,50 +1,0 @@
-templates:
-  - from: templates/generic.tmpl
-    render:
-      - to: ../prow/jobs/test-infra/test-infra-governance.yaml
-        localSets:
-          jobConfig_default:
-            command: "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/governance.sh"
-            labels:
-              preset-dind-enabled: "true"
-            request_memory: "10Mi"
-            request_cpu: "100m"
-        jobConfigs:
-          - repoName: "github.com/kyma-project/test-infra"
-            jobs:
-              - jobConfig:
-                  name: "pre-master-test-infra-governance"
-                  labels:
-                    preset-build-pr: "true"
-                  run_if_changed: "milv.config.yaml|.md$"
-                  args:
-                    - "--repository"
-                    - "test-infra"
-                inheritedConfigs:
-                  global:
-                    - "jobConfig_default"
-                    - "image_bootstrap"
-                    - "jobConfig_presubmit"
-                  local:
-                    - "jobConfig_default"
-              - jobConfig:
-                  name: "test-infra-governance-nightly"
-                  type_periodic: "true"
-                  annotations:
-                    testgrid-create-test-group: "false"
-                  cron: "0 5 * * 1-5" # “At 01:00 on every day-of-week from Monday through Friday.”
-                  cluster: "untrusted-workload"
-                  args:
-                    - "--repository"
-                    - "test-infra"
-                    - "--repository-dir"
-                    - "/home/prow/go/src/github.com/kyma-project/test-infra"
-                    - "--full-validation"
-                    - "true"
-                inheritedConfigs:
-                  global:
-                    - "jobConfig_default"
-                    - "image_bootstrap"
-                    - "extra_refs_test-infra"
-                  local:
-                    - "jobConfig_default"
